Switch to tracking in Bytes. Add client formatting for different representations.

+++ tracker.sh
@@ -4,11 +4,13 @@
 LAN_TYPE=$(nvram get lan_ipaddr | awk ' { FS="."; print $1"."$2 }')
 
 # Restore Usage from Back-up
-cat /jffs/traffic.bk | sed 's/{usage:://' | sed 's/;}//' | awk '{split($0, a, ";")} END { for(i in a) printf "%s\n", a[i] }' | awk -F':' ' {split($1, ip, ".")} {printf "export local_%s_upload=%s\nexport local_%s_download=%s\n",ip[4],$2,ip[4],$3}' > /tmp/traffic_bk.sh
-. /tmp/traffic_bk.sh
-rm /tmp/traffic_bk.sh
+if [ -f /jffs/traffic.bk ]; then
+  cat /jffs/traffic.bk | sed 's/{usage:://' | sed 's/;}//' | awk '{split($0, a, ";")} END { for(i in a) printf "%s\n", a[i] }' | awk -F':' ' {split($1, ip, ".")} {printf "export local_%s_upload=%s\nexport local_%s_download=%s\n",ip[4],$2,ip[4],$3}' > /tmp/traffic_bk.sh
+  . /tmp/traffic_bk.sh
+  rm /tmp/traffic_bk.sh
 
-mv -f /jffs/traffic.bk /tmp/www/traffic.asp
+  mv -f /jffs/traffic.bk /tmp/www/traffic.asp
+fi
 
 while :
 do
@@ -36,8 +38,8 @@ do
     fi
   done
 
-  echo "calc () { echo \$1 \$2 | awk '{ printf \"%f\", \$1 + \$2 }'; }" > /tmp/traffic_update.sh
-  iptables -L RRDIPT -vnx -t filter | grep ${LAN_TYPE} | awk '{ if ($8 == "0.0.0.0/0") { download[$9]=$2 } else if ($9 == "0.0.0.0/0") upload[$8]=$2 } END { for(item in upload) { {split(item, ip, ".")} { printf "if [ -z ${local_%s_upload+x} ]; then\necho \"%s\:%.2f\:%.2f;\";\nelse\na=`calc $local_%s_upload %.2f`;\nb=`calc $local_%s_download %.2f`;\necho \"%s\:$a\:$b;\"\nfi\n", ip[4], item, upload[item]/1048576, download[item]/1048576, ip[4], upload[item]/1048576, ip[4], download[item]/1048576, item}}}' >> /tmp/traffic_update.sh
+  echo "add () { echo \$1 \$2 | awk '{ printf \"%.0f\", \$1 + \$2 }'; }" > /tmp/traffic_update.sh
+  iptables -L RRDIPT -vnx -t filter | grep ${LAN_TYPE} | awk '{ if ($8 == "0.0.0.0/0") { download[$9]=$2 } else if ($9 == "0.0.0.0/0") upload[$8]=$2 } END { for(item in upload) { {split(item, ip, ".")} { printf "if [ -z ${local_%s_upload+x} ]; then\necho \"%s\:%.0f\:%.0f;\";\nelse\na=`add $local_%s_upload %.0f`;\nb=`add $local_%s_download %.0f`;\necho \"%s\:$a\:$b;\"\nfi\n", ip[4], item, upload[item], download[item], ip[4], upload[item], ip[4], download[item], item}}}' >> /tmp/traffic_update.sh
 
   echo "{usage::" > /tmp/traffic.dat
   . /tmp/traffic_update.sh >> /tmp/traffic.dat
